Fixed ARM deploy button.
The ARM deploy button in "Before the HOL - App modernization.md" linked an outdated/wrong version of the ARM template from a different repository (https://raw.githubusercontent.com/microsoft/MCW-App-modernization) instead of the correct one from this repository. The outdated template was missing the APIM resource leading to problems in Exercise 9 of the hands on lab. 

The link was fixed.